invoice distribution fixes
added fetch schedules endpoint

added fetch schedules endpoint

updated fetch invoice distributions

added missing/wrong functionality

+++ src/main/kotlin/com/cynergisuite/middleware/accounting/account/payable/recurring/distribution/infrastructure/AccountPayableRecurringInvoiceDistributionEntity.kt
@@ -8,6 +8,6 @@ data class AccountPayableRecurringInvoiceDistributionEntity(
    val id: UUID? = null,
    val invoiceId: UUID,
    val accountId: UUID,
-   val profitCenter: Long,
+   val profitCenter: Int,
    val amount: BigDecimal
 )
